Add new shader, plane mesh, and data fetching functionality
