+++ zarrvizSVELTE/src/lib/components/3DVolumetric/initScene.ts
@@ -1,12 +1,15 @@
 // import { writable } from 'svelte/store';
-import * as THREE from 'three';
-// import TWEEN, { Tween } from '@tweenjs/tween.js';
 import CameraControls from 'camera-controls';
-
+import * as THREE from 'three';
+// import fragmentShaderVolume from '$lib/shaders/volume.frag';
 import { getBoxSize } from '$lib/utils/Utils';
 // import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls';
 
-import { addMap } from './addMap';
+
+import fragmentShaderVolume from '$lib/shaders/simple.frag';
+import vertexShaderVolume from '$lib/shaders/volume.vert';
+
+import { addPlainMap } from './addPlainMap';
 
 // import OrbitUnlimitedControls from '@janelia/three-orbit-unlimited-controls';
 
@@ -21,6 +24,60 @@ const clock = new THREE.Clock();
 
 CameraControls.install({ THREE: THREE });
 
+
+function initMaterial(renderer, box, boxSize, sunLight, hemisphereLight,volumeDataUint8, volumeSize, voxelSize, cameraNear, cameraFar) {
+	const volumeTexture = new THREE.Data3DTexture(volumeDataUint8, volumeSize[0], volumeSize[1], volumeSize[2]);
+	volumeTexture.format = THREE.RedFormat;
+	volumeTexture.type = THREE.UnsignedByteType;
+	// Disabling mimpaps saves memory.
+	volumeTexture.generateMipmaps = false;
+	// Linear filtering disables LODs, which do not help with volume rendering.
+	volumeTexture.minFilter = THREE.LinearFilter;
+	volumeTexture.magFilter = THREE.LinearFilter;
+	volumeTexture.needsUpdate = true;
+
+	const lightColor = sunLight.color;
+	const lightColorV = new THREE.Vector3(lightColor.r, lightColor.g, lightColor.b);
+	//      const ambientLightColorV = new THREE.Vector3(0.3, 0.7, 0.98);
+	const ambientLightColorV = new THREE.Vector3(
+		hemisphereLight.color.r,
+		hemisphereLight.color.g,
+		hemisphereLight.color.b
+	);
+
+	const boxMaterial = new THREE.ShaderMaterial({
+		vertexShader: vertexShaderVolume,
+		fragmentShader: fragmentShaderVolume,
+		side: THREE.BackSide,
+		transparent: true,
+		opacity: 1.0,
+		uniforms: {
+			boxSize: new THREE.Uniform(boxSize),
+			volumeTex: new THREE.Uniform(volumeTexture),
+			voxelSize: new THREE.Uniform(voxelSize),
+			sunLightDir: new THREE.Uniform(sunLight.position),
+			sunLightColor: new THREE.Uniform(lightColorV),
+			ambientLightColor: new THREE.Uniform(ambientLightColorV),
+			near: new THREE.Uniform(cameraNear),
+			far: new THREE.Uniform(cameraFar),
+			// The following are set separately, since they are based on `props` values that can
+			// change often, and should not trigger complete re-initialization.
+			transferTex: new THREE.Uniform(null),
+			dtScale: new THREE.Uniform(0),
+			inScatFactor: new THREE.Uniform(0),
+			qLScale: new THREE.Uniform(0),
+			gHG: new THREE.Uniform(0),
+			dataEpsilon: new THREE.Uniform(0),
+			bottomColor: new THREE.Uniform(new THREE.Vector3(0.0, 0.0005, 0.0033)),
+			finalGamma: new THREE.Uniform(0)
+		}
+	});
+
+	/* eslint no-param-reassign: ["error", { "props": false }] */
+	box.material = boxMaterial;
+
+}
+
 export function initScene(
 	canvas,
 	volumeSize,
@@ -29,8 +86,10 @@ export function initScene(
 	cameraUp,
 	cameraFovDegrees,
 	cameraNear,
-	cameraFar
+	cameraFar,
+	volumeDataUint8
 ) {
+
 	const size = new THREE.Vector2();
 
 	renderer = new THREE.WebGLRenderer({
@@ -62,7 +121,9 @@ export function initScene(
 	gridHelper.position.y = -1;
 	scene.add(gridHelper);
 
-	const [boxWidth, boxHeight, boxDepth] = getBoxSize(volumeSize, voxelSize);
+
+
+	const [boxWidth = 1, boxHeight=1 , boxDepth=0.3] = getBoxSize(volumeSize, voxelSize);
 	const boxSize = new THREE.Vector3(boxWidth, boxHeight, boxDepth);
 	console.log(`Voxel size ${voxelSize[0]}, ${voxelSize[1]}, ${voxelSize[2]}`);
 	console.log(`Box size ${boxWidth}, ${boxHeight}, ${boxDepth}`);
@@ -71,8 +132,8 @@ export function initScene(
 	const box = new THREE.Mesh(boxGeometry);
 	scene.add(box);
 
-	// Lights, to be used both during rendering the volume, and rendering the optional surface.
 
+	// Lights, to be used both during rendering the volume, and rendering the optional surface.
 	const sunLightDir = new THREE.Vector3(0.0, 0.5, 0.5);
 	const sunLightColor = new THREE.Color(0.99, 0.83, 0.62);
 	const sunLight = new THREE.DirectionalLight(sunLightColor.getHex(), 1.0);
@@ -89,10 +150,8 @@ export function initScene(
 	const axesHelper = new THREE.AxesHelper(5);
 	scene.add(axesHelper);
 
-	//
-	// Add textured map to the scene
-	//
-	addMap(boxWidth, boxHeight, boxDepth);
+
+	initMaterial(renderer, box, boxSize, sunLight, hemisphereLight, volumeDataUint8, volumeSize, voxelSize, cameraNear, cameraFar);
 }
 
 export function setCameraView(position: number[], up: number[]) {
@@ -100,20 +159,7 @@ export function setCameraView(position: number[], up: number[]) {
 	camera.position.set(...position);
 	camera.up.set(...up);
 	camera.lookAt(0, 0, 0);
-	//  renderScene();
 
-	// const duration = 1000; // Duration of the animation in milliseconds
-
-	// // new Tween(camera.position)
-	// // 	.to({ x: position[0], y: position[1], z: position[2] }, duration)
-	// // 	.start();
-
-	// new Tween(camera.position)
-	// 	.to({ x: position[0], y: position[1], z: position[2] }, duration)
-	// 	.onUpdate(() => camera.lookAt(lookAt[0], lookAt[1], lookAt[2]))
-	// 	.start();
-
-	// animate();
 }
 
 function animate(time?: number) {
