+++ zarrvizSVELTE/src/lib/shaders/simple.frag
@@ -1,0 +1,5 @@
+// simple shader that paint all pixels the same color in yellow
+
+void main() {
+  gl_FragColor = vec4(1.0, 1.0, 0.0, 1.0);
+}
